fastapi/requirements.txt
fastapi
asyncpg
uvicorn
python-dotenv
mistralai